+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * enhancements
   * added **UsePersistentConnection** option to re-use the same connection for multiple queries. (by acorkery)
+  * added **ReadOnly** option to open the file in readonly mode